Adds initial Vagrantfile and ansible playbook
The Vagrantfile and the roles were inspired by https://github.com/pulp/devel/. User can
run 'vagrant up' inside the root of this repository to provision an selinuxgame box. Vagrant
take a cloud image of Fedora 25, updates all the packages, and installs a few tools that are
helpful when working with Linux. Vagrant uses the .dnf-cache directory in this repo to cache
packages between each provision of a box.